Refactor modularize project
